lib generator fixes & documentation (#6)

+++ tools/create-lib/src/generate.ts
@@ -7,22 +7,21 @@ import { fileURLToPath } from "node:url"
 import type { GenerateFunction } from "./types/GenerateFunction"
 import type { CreateLibOptions } from "./types/CreateLibOptions"
 import { argv } from "node:process"
+import { processArguments, ArgumentError } from "./utils/args"
 
 const __filename = fileURLToPath(import.meta.url)
 const __dirname = path.dirname(__filename)
-class ArgumentError extends Error {}
 
 await main()
 
 interface FileContent {
   fileName: string
   content: string
-  subFolder?: string
 }
 
 async function main() {
   try {
-    const options = getOptionsFromArgs()
+    const options = processArguments(argv)
     const templates = await getTemplatesFromFolder(options)
 
     createLibFromTemplateContent(templates, options)
@@ -35,42 +34,8 @@ async function main() {
   }
 }
 
-function getOptionsWithDefaults(libName: string) {
-  return {
-    libName,
-    packageName: `@libs/${toKebabCase(libName)}`,
-    mainFileName: "src/exports.ts",
-    templateFolder: "template",
-    isPrivate: true,
-    version: "0.0.0",
-    libFolderName: toKebabCase(libName),
-    libFunctionName: toLowerCamelCase(libName),
-  } satisfies CreateLibOptions
-}
-
-function getOptionsFromArgs(): Required<CreateLibOptions> {
-  const args = argv.slice(2)
-  if (args.length === 0) {
-    throw new ArgumentError("No arguments provided. Expected: <libName> [description] [version] [public|private]")
-  }
-  const libName = args[0]
-  const description = args[1]
-  const version = args[2]
-  const isPrivate = args[3] !== "public"
-
-  if (version && !/[^\d+.\d+.\d+$]/.test(version)) {
-    throw new ArgumentError("Version (argument 3) must be in the format of x.x.x")
-  }
-
-  return {
-    ...getOptionsWithDefaults(libName),
-    description,
-    version,
-    isPrivate,
-  }
-}
-
 type Formatter = (options: CreateLibOptions, filePath: URL) => Promise<string>
+
 async function getTemplatesFromFolder(
   options: CreateLibOptions,
   folderName?: string,
@@ -95,9 +60,10 @@ async function getTemplatesFromFolder(
       if (await isFileDirectory(filePath)) {
         return await getTemplatesFromFolder(options, fileLocation, depth + 1)
       }
+      const templatedFile = await generateTemplateFromFile(options, formatters, modulePath, fileLocation)
+
       return {
-        ...(await generateTemplateFromFile(options, formatters, modulePath, fileLocation)),
-        subFolder: folderName,
+        ...templatedFile,
         depth,
       }
     }),
@@ -117,25 +83,26 @@ async function isFileDirectory(filePath: string): Promise<boolean> {
   return stat.isDirectory()
 }
 
+function matchesExtension(fileName: string, extension: string | RegExp): string | null {
+  if (extension instanceof RegExp) {
+    const match = fileName.match(extension)
+    if (match) {
+      return match[0]
+    }
+  } else if (fileName.endsWith(extension)) {
+    return extension
+  }
+  return null
+}
+
 async function generateTemplateFromFile(
   options: CreateLibOptions,
   formatters: [string | RegExp, Formatter][],
   modulePath: URL,
   fileName: string,
 ): Promise<FileContent> {
-  const matchesExtension = (extension: string | RegExp) => {
-    if (extension instanceof RegExp) {
-      const match = fileName.match(extension)
-      if (match) {
-        return match[0]
-      }
-    } else if (fileName.endsWith(extension)) {
-      return extension
-    }
-    return null
-  }
   for (const [extension, formatter] of formatters) {
-    const extensionMatch = matchesExtension(extension)
+    const extensionMatch = matchesExtension(fileName, extension)
     if (extensionMatch) {
       return {
         fileName: formatFileName(fileName, options, extensionMatch),
@@ -177,6 +144,11 @@ function createLibFromTemplateContent(content: FileContent[], options: Required<
   for (const { fileName, content: fileContent } of content) {
     const filePath = path.join(getWorkspaceRoot(), "libs", options.libFolderName, fileName)
     const folder = path.dirname(filePath)
+    if (options.isDryRun) {
+      console.log(`Dry run: Create folder ${folder}`)
+      console.log(`Dry run: Create file ${filePath}`)
+      continue
+    }
     fs.mkdirSync(folder, { recursive: true })
     fs.writeFileSync(filePath, fileContent)
   }
@@ -184,17 +156,4 @@ function createLibFromTemplateContent(content: FileContent[], options: Required<
 
 function getWorkspaceRoot() {
   return path.join(__dirname, "..", "..", "..")
-}
-
-function toKebabCase(string: string): string {
-  return string
-    .replace(/[\s_]+/g, "-")
-    .toLowerCase()
-    .replace(/[^a-z0-9-]/g, "")
-}
-
-function toLowerCamelCase(string: string): string {
-  return string
-    .replace(/(?:^\w|[A-Z]|\b\w)/g, (letter, index) => (index === 0 ? letter.toLowerCase() : letter.toUpperCase()))
-    .replace(/[^a-zA-Z0-9_]/g, "")
 }
